Contrato em template, visualizar e PDF
