+++ src/helpers/PDFGenerator/index.tsx
@@ -1,102 +1,96 @@
 import jsPDF from "jspdf";
 
 const PdfGenerator = (
-  documentContent: Document,
-  elementId: string,
-  fileName: string
+    documentContent: Document,
+    elementId: string,
+    fileName: string,
+    output: "window" | "pdf"
 ): void => {
-  const element = documentContent.getElementById(elementId);
-  if (element) {
-    const elemento = document.getElementById(elementId);
-    const htmlContent = elemento ? elemento.innerHTML : "";
+    const element = documentContent.getElementById(elementId);
 
-    const estilo = `
-            body, p, span {
-                font-family:  Roboto, sans-serif;
-                font-weight: 300;
-                font-size: 10px;
-                margin: 0;
-                padding: 0;
+    if (element) {
+        const htmlContent = element.innerHTML;
+
+        if (output === "pdf") {
+            // **Gerar e salvar o arquivo PDF**
+            const pdf = new jsPDF("p", "pt", "a4");
+            const marginX = 20;
+            const marginY = 20;
+            const pageWidth = pdf.internal.pageSize.getWidth() - marginX * 2;
+            const pageHeight = pdf.internal.pageSize.getHeight() - marginY * 2;
+            const elementWidth = element.offsetWidth;
+            const elementHeight = element.offsetHeight;
+            const scaleX = pageWidth / elementWidth;
+            const scaleY = pageHeight / elementHeight;
+            const scale = Math.min(scaleX, scaleY);
+
+            pdf.html(element, {
+                callback: function (pdf) {
+                    pdf.save(fileName);
+                },
+                x: marginX,
+                y: marginY,
+                html2canvas: {
+                    scale: scale,
+                },
+            });
+        } else if (output === "window") {
+            // **Exibir o HTML formatado em uma nova aba do navegador, sem sugestão de impressão**
+            const estilo = `
+                <style>
+                    body {
+                      font-family: 'Roboto', sans-serif;
+                      font-weight: 400;
+                      font-size: 16px;
+                      letter-spacing: .6px;
+                      word-spacing: 0px;
+                      line-height: 1.2;
+                      margin: 50px;
+                    }
+                    p {
+                        margin: 0;
+                    }
+                </style>
+            `;
+            // const estilo = `
+            //     <style>
+            //         body {
+            //             font-family: Roboto, sans-serif;
+            //             font-weight: 400;
+            //             font-size: 16px;
+            //             margin: 0;
+            //             padding: 0;
+            //             line-height: 1.2; /* Ajuste do espaçamento entre linhas */
+            //         }
+            //         p {
+            //             margin-bottom: 10px; /* Ajuste do espaço entre parágrafos */
+            //         }
+            //     </style>
+            // `;
+
+            //
+
+            const newWindow = window.open("", "_blank");
+            if (newWindow) {
+                newWindow.document.write(`
+                    <html>
+                        <head>
+                            <title>${fileName}</title>
+                            ${estilo}
+                        </head>
+                        <body>${htmlContent}</body>
+                    </html>
+                `);
+                newWindow.document.close(); // Fecha o stream de escrita e mostra o conteúdo
+            } else {
+                console.error("Falha ao abrir uma nova janela.");
             }
-        `;
-
-    const iframe = document.createElement("iframe");
-    iframe.style.position = "absolute";
-    iframe.style.border = "none";
-
-    document.body.appendChild(iframe);
-
-    const iframeWindow = iframe.contentWindow;
-    if (!iframeWindow) {
-      console.log("Falha ao acessar o contentWindow do iframe.");
-      return;
+        }
+    } else {
+        console.error(
+            `PDFGenerator: Element with ID '${elementId}' not found.`
+        );
     }
-
-    const doc = iframeWindow.document;
-    doc.open();
-    doc.write(`
-            <html>
-                <head>
-                    <style>
-                        ${estilo}
-                    </style>
-                </head>
-                <body>
-                    ${htmlContent}
-                </body>
-            </html>
-        `);
-    doc.close();
-
-    iframe.onload = () => {
-      const imgs = doc.getElementsByTagName("img");
-      let loadedImages = 0;
-
-      const checkAllImagesLoaded = () => {
-        if (loadedImages === imgs.length) {
-          iframeWindow.focus();
-          //iframeWindow.print(); // Imprimir pelo windows
-
-          const pdf = new jsPDF("p", "pt", "a4");
-          const marginX = 20;
-          const marginY = 20;
-          const pageWidth = pdf.internal.pageSize.getWidth() - marginX * 2;
-          const pageHeight = pdf.internal.pageSize.getHeight() - marginY * 2;
-          const elementWidth = element.offsetWidth;
-          const elementHeight = element.offsetHeight;
-          const scaleX = pageWidth / elementWidth;
-          const scaleY = pageHeight / elementHeight;
-          const scale = Math.min(scaleX, scaleY);
-
-          pdf.html(element, {
-            callback: function (pdf) {
-              pdf.save(fileName);
-            },
-            x: marginX,
-            y: marginY,
-            html2canvas: {
-              scale: scale,
-            },
-          });
-
-          document.body.removeChild(iframe);
-        }
-      };
-
-      if (imgs.length === 0) {
-        iframeWindow.focus();
-        document.body.removeChild(iframe);
-      } else {
-        for (let img of imgs) {
-          img.src = img.src.replace("low-res", "high-res");
-          loadedImages++;
-          checkAllImagesLoaded();
-        }
-      }
-    };
-  } else {
-    console.error(`PDFGenerator: Element with ID '${elementId}' not found.`);
-  }
 };
 
 export default PdfGenerator;
